adding 0-

+++ 0-bubble_sort.c
@@ -26,7 +26,7 @@ void swap(int *a, int *b)
 */
 void bubble_sort(int *array, size_t size)
 {
-	int i = 0, j = 0;
+	size_t i = 0, j = 0;
 
 	for (; i < size - 1 ; i++)
 	{
